+++ packages/eslint-plugin-fiori-tools/eslintrc-typescript.js
@@ -17,7 +17,10 @@ const overrides = [
             '**/*.d.ts'
         ],
         'parser': '@typescript-eslint/parser',
-        'extends': ['plugin:@typescript-eslint/recommended'],
+        'extends': [
+            'plugin:@typescript-eslint/recommended',
+            'plugin:@typescript-eslint/recommended-requiring-type-checking'
+        ],
         'parserOptions': {
             'project': true // uses local tsconfig.json nearest to file being linted. Especially important for monorepos
         }