+++ packages/preview-middleware-client/src/adp/init.ts
@@ -54,7 +54,7 @@ export default async function (rta: RuntimeAuthoring) {
     }
 
     // also initialize the editor
-    init(rta);
+    await init(rta);
 
     const ui5VersionValidationMsg = getUI5VersionValidationMessage(version);