+++ packages/ui5-library-writer/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sap-ux/ui5-library-writer
 
+## 0.3.29
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.19.8
+-   @sap-ux/ui5-application-writer@0.26.7
+
+## 0.3.28
+
+### Patch Changes
+
+-   Updated dependencies [98496d57]
+-   Updated dependencies [e3d2e003]
+    -   @sap-ux/project-access@1.19.7
+    -   @sap-ux/ui5-application-writer@0.26.7
+
+## 0.3.27
+
+### Patch Changes
+
+-   da0ecd9a: Enable Typscript type checking in eslint module @sap-ux/eslint-plugin-fiori-tools
+-   Updated dependencies [da0ecd9a]
+    -   @sap-ux/ui5-application-writer@0.26.7
+
 ## 0.3.26
 
 ### Patch Changes